+++ src/lib.rs
@@ -181,6 +181,9 @@ pub mod buffer;
 pub mod conversions;
 pub mod decoder;
 pub mod math;
+#[cfg(feature = "recording")]
+/// Microphone input support for audio recording.
+pub mod microphone;
 pub mod mixer;
 pub mod queue;
 pub mod source;